+++ src/background/modules/monitors/application.rs
@@ -163,8 +163,8 @@ impl MonitorManager {
 
     fn get_monitors() -> Result<Vec<(String, HMONITOR)>> {
         let mut monitors = Vec::new();
-        for m in MonitorEnumerator::get_all()? {
-            monitors.push((WindowsApi::monitor_name(m)?, m));
+        for m in MonitorEnumerator::get_all_v2()? {
+            monitors.push((m.device_id()?, m.raw()));
         }
         Ok(monitors)
     }